==> R/03_description.qmd <==
---
title: "03_description"
format: html
editor: visual
---

## Libraries

```{r}
#| silent: true
#| message: false

library("here")
library("tidyverse")
source("99_proj_func.R")
```

Now that we have already imported the necessary libraries, we proceed to load the files and merge them

```{r}

# load metadata
female_metadata <- read_csv(here("data","02_female_metadata.csv"))
male_metadata <- read_csv(here("data", "02_male_metadata.csv"))

# load normalised data
female_norm_counts <- read_csv(here("data","02_norm_counts_filtered_female.csv"))
male_norm_counts <- read_csv(here("data", "02_norm_counts_filtered_male.csv"))

# load raw counts data 
counts_filtered_female <- read_csv(here("data","02_counts_filtered_female.csv"))
counts_filtered_male <- read_csv(here("data","02_counts_filtered_male.csv"))
```

## Age

Plot age distribution

```{r}

female_bp <- ggplot(data = female_metadata,
       map = aes(x = cancer_status,
                 y = age_at_diag,
                 fill = cancer_status)) +
  geom_boxplot() +
  labs(title = "Age distribution of female patients",
       x = "cancer status",
       y = "age at diagnosis") +
    theme_minimal()

male_bp <- ggplot(data = male_metadata,
       map = aes(x = cancer_status,
                 y = age_at_diag)) +
  geom_boxplot() +
  labs(title = "Age distribution of male patients",
       x = "cancer status",
       y = "age at diagnosis") +
    theme_minimal()


ggsave(
  here("results", "03_female_age_distribution.png"), 
  plot =female_bp, 
  width = 5, 
  height = 6, 
  dpi = 300,
  bg = "white")


ggsave(
  here("results", "03_male_age_distribution.png"), 
  plot =male_bp, 
  width = 5, 
  height = 6, 
  dpi = 300,
  bg = "white")
```

## Country

Plot the country of our patients distribution:

```{r}

bar_country <- ggplot(data = female_metadata,
       map = aes(x = country,
                 fill = country)) +
  geom_bar() +
  labs(title = "Origin of female patients") +
  theme_minimal()

ggsave(
  here("results", "03_female_country_distribution.png"), 
  plot =bar_country, 
  width = 10, 
  height = 6, 
  dpi = 300,
  bg = "white")
```

```{r}
bar_country_male <- ggplot(data = male_metadata,
       map = aes(x = country,
              fill = country)) +
  geom_bar() +
  labs(title = "Origin of male patients") +
  theme_minimal()

ggsave(
  here("results", "03_male_country_distribution.png"), 
  plot =bar_country_male, 
  width = 10, 
  height = 6, 
  dpi = 300,
  bg = "white")
```

## Cancer Types

Plot the distribution of cancer types in our data set:

```{r}

sample_histo <- ggplot(data = female_metadata,
       map = aes(x = histol_type,
                 fill = sample_type)) +
  geom_bar(position = "dodge") + 
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) +
  labs(title = "Distribution of Histological types",
       x = "histol type",
       fill = "sample type")


ggsave(
  here("results", "03_sample_histo_barplot.png"), 
  plot =sample_histo, 
  width = 10, 
  height = 6, 
  dpi = 300,
  bg = "white")

```
